+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/util/PreferencesUtils.java
@@ -81,6 +81,11 @@ public class PreferencesUtils {
         return sharedPreferences.getInt(keyId, defaultValue);
     }
 
+    @SuppressWarnings("SameParameterValue")
+    public static int getInt(Context context, int keyId, int defaultValue) {
+        return getInt(context, getKey(context, keyId), defaultValue);
+    }
+
     public static void setInt(Context context, String keyId, int value) {
         SharedPreferences sharedPreferences = PreferenceManager
                 .getDefaultSharedPreferences(context);